kdb: Fix printf format warning for phys_addr_t
When building without PAE support, phys_addr_t is defined as unsigned long,
but the kdb printf call uses %llx, which expects an unsigned long long.
This triggers a -Wformat warning due to a type mismatch.

Fix this by explicitly casting the phys_addr_t value to unsigned long long,
ensuring the format string and argument type always match.

This avoids build warnings while preserving existing type definitions.
Message-ID: <20250624013544.842-1-nikic.milos@gmail.com>

+++ i386/i386at/acpi_parse_apic.c
@@ -48,7 +48,7 @@ acpi_print_info(phys_addr_t rsdp, void *rsdt, int acpi_rsdt_n)
 {
 
     printf("ACPI:\n");
-    printf(" rsdp = 0x%llx\n", rsdp);
+    printf(" rsdp = 0x%llx\n", (unsigned long long) rsdp);
     printf(" rsdt/xsdt = 0x%p (n = %d)\n", rsdt, acpi_rsdt_n);
 }
